+++ stdlib/lib.core.d.ts
@@ -966,6 +966,7 @@ interface FSCpOptions {
 }
 type FSEncoding = "utf8" | "utf-8";
 type FSReadFileStringEncoding = FSEncoding | "hex" | "base64";
+type FSFileContentEncoding = FSEncoding | "hex" | "base64";
 type FSBufferEncoding = "buffer";
 interface FSEncodingOptions {
     encoding?: FSEncoding;
@@ -984,17 +985,21 @@ interface FSReadFileBufferObjectOptions extends FSBufferEncodingOptions {
 type FSReadFileBufferOptions = FSBufferEncoding | null | FSReadFileBufferObjectOptions;
 type FSReadFileStringOptions = FSReadFileStringEncoding | FSReadFileOptions;
 type FSWriteFileFlag = "w" | "wx" | "w+" | "wx+" | "a" | "ax" | "a+" | "ax+" | "as" | "as+" | "r+" | "rs+";
-interface FSWriteFileOptions extends FSEncodingOptions {
+interface FSWriteFileOptions {
+    encoding?: FSFileContentEncoding;
     flag?: FSWriteFileFlag;
     mode?: number;
     flush?: boolean;
 }
+type FSWriteFileEncodingOptions = FSFileContentEncoding | FSWriteFileOptions;
 type FSAppendFileFlag = "a" | "ax" | "a+" | "ax+" | "as" | "as+";
-interface FSAppendFileOptions extends FSEncodingOptions {
+interface FSAppendFileOptions {
+    encoding?: FSFileContentEncoding;
     flag?: FSAppendFileFlag;
     mode?: number;
     flush?: boolean;
 }
+type FSAppendFileEncodingOptions = FSFileContentEncoding | FSAppendFileOptions;
 interface FSReaddirOptions extends FSEncodingOptions {
     recursive?: boolean;
     withFileTypes?: false;
@@ -1025,8 +1030,8 @@ interface FS {
     readonly constants: FSConstants;
     readFileSync(path: FSPathLike, options: FSReadFileBufferOptions, ...ignored: any[]): Buffer;
     readFileSync(path: FSPathLike, options?: FSReadFileStringOptions, ...ignored: any[]): string;
-    writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSWriteFileOptions, ...ignored: any[]): void;
-    appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSAppendFileOptions, ...ignored: any[]): void;
+    writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSWriteFileEncodingOptions, ...ignored: any[]): void;
+    appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSAppendFileEncodingOptions, ...ignored: any[]): void;
     existsSync(path: FSPathLike, ...ignored: any[]): boolean;
     accessSync(path: FSPathLike, mode?: number, ...ignored: any[]): void;
     readdirSync(path: FSPathLike, options: FSBufferEncoding | FSReaddirBufferOptions, ...ignored: any[]): Buffer[];
@@ -1062,8 +1067,8 @@ interface FS {
 interface FSPromises {
     readFile(path: FSPathLike, options: FSReadFileBufferOptions, ...ignored: any[]): Promise<Buffer>;
     readFile(path: FSPathLike, options?: FSReadFileStringOptions, ...ignored: any[]): Promise<string>;
-    writeFile(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSWriteFileOptions, ...ignored: any[]): Promise<void>;
-    appendFile(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSAppendFileOptions, ...ignored: any[]): Promise<void>;
+    writeFile(path: FSPathLike, data: string | Buffer, options?: FSWriteFileEncodingOptions, ...ignored: any[]): Promise<void>;
+    appendFile(path: FSPathLike, data: string | Buffer, options?: FSAppendFileEncodingOptions, ...ignored: any[]): Promise<void>;
     readdir(path: FSPathLike, options: FSBufferEncoding | FSReaddirBufferOptions, ...ignored: any[]): Promise<Buffer[]>;
     readdir(path: FSPathLike, options: FSReaddirDirentOptions, ...ignored: any[]): Promise<FSDirent[]>;
     readdir(path: FSPathLike, options?: FSFileEncodingOptions | FSReaddirOptions, ...ignored: any[]): Promise<string[]>;
@@ -1100,8 +1105,8 @@ declare module "fs" {
     export const promises: FSPromises;
     export function readFileSync(path: FSPathLike, options: FSReadFileBufferOptions, ...ignored: any[]): Buffer;
     export function readFileSync(path: FSPathLike, options?: FSReadFileStringOptions, ...ignored: any[]): string;
-    export function writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSWriteFileOptions, ...ignored: any[]): void;
-    export function appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSAppendFileOptions, ...ignored: any[]): void;
+    export function writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSWriteFileEncodingOptions, ...ignored: any[]): void;
+    export function appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSAppendFileEncodingOptions, ...ignored: any[]): void;
     export function existsSync(path: FSPathLike, ...ignored: any[]): boolean;
     export function accessSync(path: FSPathLike, mode?: number, ...ignored: any[]): void;
     export function readdirSync(path: FSPathLike, options: FSBufferEncoding | FSReaddirBufferOptions, ...ignored: any[]): Buffer[];
@@ -1138,8 +1143,8 @@ declare module "node:fs" {
     export const promises: FSPromises;
     export function readFileSync(path: FSPathLike, options: FSReadFileBufferOptions, ...ignored: any[]): Buffer;
     export function readFileSync(path: FSPathLike, options?: FSReadFileStringOptions, ...ignored: any[]): string;
-    export function writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSWriteFileOptions, ...ignored: any[]): void;
-    export function appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSFileEncodingOptions | FSAppendFileOptions, ...ignored: any[]): void;
+    export function writeFileSync(path: FSPathLike, data: string | Buffer, options?: FSWriteFileEncodingOptions, ...ignored: any[]): void;
+    export function appendFileSync(path: FSPathLike, data: string | Buffer, options?: FSAppendFileEncodingOptions, ...ignored: any[]): void;
     export function existsSync(path: FSPathLike, ...ignored: any[]): boolean;
     export function accessSync(path: FSPathLike, mode?: number, ...ignored: any[]): void;
     export function readdirSync(path: FSPathLike, options: FSBufferEncoding | FSReaddirBufferOptions, ...ignored: any[]): Buffer[];